feat(params): make assemblies first-class parameter holders (#1557)
Assemblies own a parameter table (AssemblyComponentDefinition.params) and
share it into their sketches, but the assembly recipe never persisted it: every
user/text/boolean parameter and any parameter-driven assembly sketch dimension
was silently lost on save. M39-F01 closes that gap.

- Extract the parameter-recipe codec out of the part definition into a shared,
  table-keyed module (param_recipe.go) so the part and assembly recipes use one
  conversion instead of duplicating ~150 lines.
- Persist Parameters/Groups/Settings/DerivedTables in assemblyRecipe and restore
  them before the assembly sketches, so a dimension expression rebinds to its
  parameter by name on reload.
- Wire ShareParameters on the assembly sketch collection (constructor + the
  snapshot-restore reset) so restored sketches bind dimension expressions against
  the populated table as their dimensions rebuild — the same correct-by-
  construction timing the part already relies on.
- Reset the parameter table in resetOccurrences so an undo/redo snapshot restore
  is a true full replace, not a merge.

Foundation for M39 F02-F04 (assembly<->part parameter linking).
