+++ model/compdef/serialize.go
@@ -199,10 +199,10 @@ func (d *PartComponentDefinition) buildRecipe() (partRecipe, error) {
 		}
 	}
 	r.Units = d.unitsRecipe()
-	r.Parameters = d.parametersRecipe()
-	r.ParameterGroups = d.parameterGroupsRecipe()
-	r.ParameterSettings = d.parameterSettingsRecipe()
-	r.DerivedTables = d.derivedTablesRecipe()
+	r.Parameters = parametersRecipeOf(d.params)
+	r.ParameterGroups = parameterGroupsRecipeOf(d.params)
+	r.ParameterSettings = parameterSettingsRecipeOf(d.params)
+	r.DerivedTables = derivedTablesRecipeOf(d.params)
 	r.Materials = d.materialsRecipe()
 	r.Properties = propertiesRecipeOf(d.props)
 	r.SheetMetal = d.sheetMetalRecipeOf()
@@ -252,13 +252,13 @@ func (d *PartComponentDefinition) applyRecipeStruct(r partRecipe) error {
 		return err
 	}
 	applyPropertiesRecipe(d.props, r.Properties)
-	if err := d.applyParameters(r.ParameterGroups, r.Parameters); err != nil {
+	if err := applyParametersTo(d.params, r.ParameterGroups, r.Parameters); err != nil {
 		return err
 	}
-	if err := d.applyParameterSettings(r.ParameterSettings); err != nil {
+	if err := applyParameterSettingsTo(d.params, r.ParameterSettings); err != nil {
 		return err
 	}
-	if err := d.applyDerivedTables(r.DerivedTables); err != nil {
+	if err := applyDerivedTablesTo(d.params, r.DerivedTables); err != nil {
 		return err
 	}
 	if err := feature.ApplyWork(d.work, r.WorkFeatures); err != nil {
@@ -423,315 +423,6 @@ func parseAngleFormat(s string) (param.AngleFormat, bool) {
 		return param.AngleDMS, true
 	}
 	return param.AngleDecimal, false
-}
-
-// parametersRecipe captures every parameter in creation order (a valid order to
-// re-add: an expression can only reference parameters created before it).
-func (d *PartComponentDefinition) parametersRecipe() []parameterRecipe {
-	all := d.params.All()
-	if len(all) == 0 {
-		return nil
-	}
-	out := make([]parameterRecipe, 0, len(all))
-	for _, p := range all {
-		out = append(out, d.parameterRecipeOf(p))
-	}
-	return out
-}
-
-// parameterRecipeOf captures one parameter's persisted form: its value (by flavor) plus
-// the shared comment/key/export/precision/tolerance/multi-value/group state.
-func (d *PartComponentDefinition) parameterRecipeOf(p *param.Parameter) parameterRecipe {
-	pr := parameterRecipe{
-		Name: p.Name(), Kind: p.Kind().String(),
-		Comment: p.Comment, Key: p.IsKey, Export: p.ExposedAsProperty, Precision: p.Precision,
-	}
-	switch {
-	case p.IsText():
-		pr.ValueType, pr.Text = "text", p.Text()
-	case p.IsBoolean():
-		pr.ValueType, pr.Bool = "boolean", p.Bool()
-	case p.Kind().Editable():
-		pr.Expression = p.Expression()
-	default:
-		pr.Value, pr.Unit = p.Value().Value, p.Value().Unit.String()
-	}
-	if t := p.Tolerance(); t != (param.Tolerance{}) {
-		pr.Tolerance = &toleranceRecipe{Type: t.Kind().String(), Upper: t.Upper, Lower: t.Lower}
-	}
-	if m := p.ModelValueType(); m != param.Nominal {
-		pr.ModelValueType = m.String()
-	}
-	if p.IsMultiValue() {
-		pr.ExprList, pr.AllowCustom = p.ExpressionList(), p.AllowsCustomValue()
-		pr.SortedValueList = !p.CustomOrder()
-	}
-	pr.Hidden = !p.Visible
-	if p.DisplayFormat != param.DisplayFormatDecimal {
-		pr.DisplayFormat = p.DisplayFormat.String()
-	}
-	if cp := p.CustomProperty; cp != param.DefaultCustomPropertyFormat() {
-		pr.CustomProperty = &customPropertyRecipe{
-			PropertyType: cp.PropertyType.String(), Units: cp.Units, Precision: cp.Precision.String(),
-			ShowLeadingZeros: cp.ShowLeadingZeros, ShowTrailingZeros: cp.ShowTrailingZeros,
-			ShowUnitsString: cp.ShowUnitsString,
-		}
-	}
-	return pr
-}
-
-// applyParameters re-creates the custom groups (in their saved order) then re-adds each
-// parameter in recipe order. A parse error (bad expression, duplicate name, unknown
-// kind/unit) aborts the load rather than dropping the parameter silently.
-func (d *PartComponentDefinition) applyParameters(groups []parameterGroupRecipe, params []parameterRecipe) error {
-	for _, pr := range params {
-		p, err := d.addParameter(pr)
-		if err != nil {
-			return fmt.Errorf("compdef: restore parameter %q: %w", pr.Name, err)
-		}
-		if err := d.applyParameterState(p, pr); err != nil {
-			return fmt.Errorf("compdef: restore parameter %q: %w", pr.Name, err)
-		}
-	}
-	// Groups restore after the parameters their member lists name.
-	for _, gr := range groups {
-		if err := d.applyParameterGroup(gr); err != nil {
-			return fmt.Errorf("compdef: restore parameter group %q: %w", gr.InternalName, err)
-		}
-	}
-	return nil
-}
-
-// derivedTablesRecipe renders the derived parameter tables in creation order.
-func (d *PartComponentDefinition) derivedTablesRecipe() []derivedTableRecipe {
-	var out []derivedTableRecipe
-	for _, t := range d.params.DerivedTables() {
-		out = append(out, derivedTableRecipe{
-			ID: t.ID(), SourceDocument: t.SourceDocument(),
-			Linked: t.Linked(), OwnedByFeature: t.OwnedByFeature(),
-		})
-	}
-	return out
-}
-
-// applyDerivedTables re-creates the persisted tables after the parameters
-// they reconnect to.
-func (d *PartComponentDefinition) applyDerivedTables(tables []derivedTableRecipe) error {
-	for _, tr := range tables {
-		if err := d.params.RestoreDerivedTable(tr.ID, tr.SourceDocument, tr.Linked, tr.OwnedByFeature); err != nil {
-			return fmt.Errorf("compdef: restore derived table %d: %w", tr.ID, err)
-		}
-	}
-	return nil
-}
-
-// parameterSettingsRecipe persists the parameter settings, nil when they are
-// the new-document defaults (the recipe omits zero state).
-func (d *PartComponentDefinition) parameterSettingsRecipe() *parameterSettingsRecipe {
-	s := *d.params.Settings()
-	if s == param.DefaultCollectionSettings() {
-		return nil
-	}
-	return &parameterSettingsRecipe{
-		LinearStandardTolerance:      s.LinearStandardTolerance,
-		AngularStandardTolerance:     s.AngularStandardTolerance,
-		UseStandardTolerances:        s.UseStandardTolerances,
-		ExportStandardTolerances:     s.ExportStandardTolerances,
-		LinearDimensionPrecision:     s.LinearDimensionPrecision,
-		AngularDimensionPrecision:    s.AngularDimensionPrecision,
-		DimensionDisplayType:         s.DimensionDisplayType.String(),
-		DisplayParameterAsExpression: s.DisplayParameterAsExpression,
-	}
-}
-
-// applyParameterSettings restores the persisted parameter settings (nil keeps
-// the defaults).
-func (d *PartComponentDefinition) applyParameterSettings(sr *parameterSettingsRecipe) error {
-	if sr == nil {
-		return nil
-	}
-	display, ok := types.ParseDimensionDisplayType(sr.DimensionDisplayType)
-	if !ok {
-		return fmt.Errorf("compdef: unknown dimension display type %q (want value|name|expression|tolerance|preciseValue)", sr.DimensionDisplayType)
-	}
-	*d.params.Settings() = param.CollectionSettings{
-		LinearStandardTolerance:      sr.LinearStandardTolerance,
-		AngularStandardTolerance:     sr.AngularStandardTolerance,
-		UseStandardTolerances:        sr.UseStandardTolerances,
-		ExportStandardTolerances:     sr.ExportStandardTolerances,
-		LinearDimensionPrecision:     sr.LinearDimensionPrecision,
-		AngularDimensionPrecision:    sr.AngularDimensionPrecision,
-		DimensionDisplayType:         display,
-		DisplayParameterAsExpression: sr.DisplayParameterAsExpression,
-	}
-	return nil
-}
-
-// parameterGroupsRecipe renders the custom groups with their member names, in
-// creation order.
-func (d *PartComponentDefinition) parameterGroupsRecipe() []parameterGroupRecipe {
-	var out []parameterGroupRecipe
-	for _, g := range d.params.Groups() {
-		gr := parameterGroupRecipe{InternalName: g.InternalName(), ClientID: g.ClientID}
-		if g.DisplayName != g.InternalName() {
-			gr.DisplayName = g.DisplayName
-		}
-		for _, id := range d.params.GroupMembers(g.InternalName()) {
-			if p, ok := d.params.ByID(id); ok {
-				gr.Members = append(gr.Members, p.Name())
-			}
-		}
-		out = append(out, gr)
-	}
-	return out
-}
-
-// applyParameterGroup re-creates one custom group and re-attaches its members
-// by name.
-func (d *PartComponentDefinition) applyParameterGroup(gr parameterGroupRecipe) error {
-	g, err := d.params.AddGroup(gr.InternalName, gr.DisplayName, gr.ClientID)
-	if err != nil {
-		return err
-	}
-	for _, name := range gr.Members {
-		p, ok := d.params.ByName(name)
-		if !ok {
-			return fmt.Errorf("member %q is not a parameter", name)
-		}
-		if err := d.params.AddToGroup(p.ID(), g.InternalName()); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// addParameter re-creates one parameter's value from its recipe entry, returning it so the
-// shared state can be applied. Read-only parameters return nil (no editable state to set).
-func (d *PartComponentDefinition) addParameter(pr parameterRecipe) (*param.Parameter, error) {
-	switch pr.ValueType {
-	case "text":
-		return d.params.AddTextUserParameter(pr.Name, pr.Text)
-	case "boolean":
-		return d.params.AddBooleanUserParameter(pr.Name, pr.Bool)
-	}
-	switch pr.Kind {
-	case param.UserParam.String():
-		return d.params.AddUserParameter(pr.Name, pr.Expression)
-	case param.ModelParam.String():
-		return d.params.AddModelParameter(pr.Name, pr.Expression)
-	case param.TableParam.String():
-		return d.params.AddTableParameter(pr.Name, pr.Expression)
-	case param.ReferenceParam.String():
-		return d.addReadOnlyParameter(pr, d.params.AddReferenceParameter)
-	case param.DerivedParam.String():
-		return d.addReadOnlyParameter(pr, d.params.AddDerivedParameter)
-	default:
-		return nil, fmt.Errorf("unknown parameter kind %q (want user|model|table|reference|derived)", pr.Kind)
-	}
-}
-
-// applyParameterState restores the shared presentation/behavior fields onto a freshly
-// re-added parameter: comment, key, export, precision, visibility, display format,
-// tolerance, multi-value list, custom-property format, group.
-func (d *PartComponentDefinition) applyParameterState(p *param.Parameter, pr parameterRecipe) error {
-	p.Comment, p.IsKey, p.ExposedAsProperty, p.Precision = pr.Comment, pr.Key, pr.Export, pr.Precision
-	p.Visible = !pr.Hidden
-	if err := applyParameterFormats(p, pr); err != nil {
-		return err
-	}
-	if err := applyParameterTolerance(p, pr); err != nil {
-		return err
-	}
-	return applyParameterValueList(p, pr)
-}
-
-// applyParameterFormats restores the display format and custom-property format
-// from their persisted wire spellings.
-func applyParameterFormats(p *param.Parameter, pr parameterRecipe) error {
-	if pr.DisplayFormat != "" {
-		f, ok := types.ParseParameterDisplayFormat(pr.DisplayFormat)
-		if !ok {
-			return fmt.Errorf("unknown display format %q (want decimal|fractional|architectural)", pr.DisplayFormat)
-		}
-		p.DisplayFormat = f
-	}
-	if pr.CustomProperty == nil {
-		return nil
-	}
-	cp, err := customPropertyFromRecipe(*pr.CustomProperty)
-	if err != nil {
-		return err
-	}
-	p.CustomProperty = cp
-	return nil
-}
-
-// customPropertyFromRecipe parses one persisted custom-property format,
-// defaulting absent enum spellings to the new-parameter defaults.
-func customPropertyFromRecipe(cr customPropertyRecipe) (param.CustomPropertyFormat, error) {
-	cp := param.DefaultCustomPropertyFormat()
-	cp.Units, cp.ShowLeadingZeros, cp.ShowTrailingZeros, cp.ShowUnitsString =
-		cr.Units, cr.ShowLeadingZeros, cr.ShowTrailingZeros, cr.ShowUnitsString
-	if cr.PropertyType != "" {
-		t, ok := types.ParseCustomPropertyType(cr.PropertyType)
-		if !ok {
-			return cp, fmt.Errorf("unknown custom property type %q (want text|number)", cr.PropertyType)
-		}
-		cp.PropertyType = t
-	}
-	if cr.Precision != "" {
-		prec, ok := types.ParseCustomPropertyPrecision(cr.Precision)
-		if !ok {
-			return cp, fmt.Errorf("unknown custom property precision %q", cr.Precision)
-		}
-		cp.Precision = prec
-	}
-	return cp, nil
-}
-
-// applyParameterTolerance restores the tolerance band and the model-value
-// selection from their persisted wire spellings.
-func applyParameterTolerance(p *param.Parameter, pr parameterRecipe) error {
-	if pr.Tolerance != nil {
-		t, ok := types.ParseToleranceType(pr.Tolerance.Type)
-		if !ok {
-			return fmt.Errorf("unknown tolerance type %q", pr.Tolerance.Type)
-		}
-		p.SetTolerance(param.Tolerance{Type: t, Upper: pr.Tolerance.Upper, Lower: pr.Tolerance.Lower})
-	}
-	if pr.ModelValueType == "" {
-		return nil
-	}
-	m, ok := types.ParseModelValueType(pr.ModelValueType)
-	if !ok {
-		return fmt.Errorf("unknown model value type %q (want nominal|lower|upper|median)", pr.ModelValueType)
-	}
-	return p.SetModelValueType(m)
-}
-
-// applyParameterValueList restores the multi-value choices, re-sorting when the
-// list was saved as auto-sorted.
-func applyParameterValueList(p *param.Parameter, pr parameterRecipe) error {
-	if len(pr.ExprList) == 0 {
-		return nil
-	}
-	if err := p.SetExpressionList(pr.ExprList, pr.AllowCustom); err != nil {
-		return err
-	}
-	if pr.SortedValueList {
-		p.SetCustomOrder(false)
-	}
-	return nil
-}
-
-// addReadOnlyParameter rebuilds a read-only parameter's measured quantity from its
-// value + unit and adds it through the given collection method.
-func (d *PartComponentDefinition) addReadOnlyParameter(pr parameterRecipe, add func(string, param.Quantity) (*param.Parameter, error)) (*param.Parameter, error) {
-	unit, ok := unitCategoryByName(pr.Unit)
-	if !ok {
-		return nil, fmt.Errorf("unknown unit %q", pr.Unit)
-	}
-	return add(pr.Name, param.Q(pr.Value, unit))
 }
 
 // unitCategoryByName maps a category name (param.Unit.String()) back to its Unit.
